Simplified and organized jacobian code.
